requirements.txt
streamlit
opencv-python-headless==4.8.0.76
tensorflow-cpu==2.12.0
numpy
pandas
protobuf==3.20.*
requests
gdown